requirements.txt
Flask-Cors==3.0.10
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-jwt-extended==4.7.1